pigmi-backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
firebase-admin>=6.0.0
python-dotenv>=1.0.0
gunicorn
